+++ core/source/org/libreoffice/ide/eclipse/core/launch/office/LaunchConfigurationTabs.java
@@ -52,10 +52,10 @@ public class LaunchConfigurationTabs extends AbstractLaunchConfigurationTabGroup
     public void createTabs(ILaunchConfigurationDialog pDialog, String pMode) {
         ILaunchConfigurationTab[] tabs = null;
         if (ILaunchManager.DEBUG_MODE.equals(pMode)) {
-            tabs = new ILaunchConfigurationTab[] { new OfficeTab(), new PackageConfigTab(), new SourceLookupTab(),
+            tabs = new ILaunchConfigurationTab[] { new OfficeTab(), new SourceLookupTab(),
                 new CommonTab() };
         } else {
-            tabs = new ILaunchConfigurationTab[] { new OfficeTab(), new PackageConfigTab(), new CommonTab() };
+            tabs = new ILaunchConfigurationTab[] { new OfficeTab(), new CommonTab() };
         }
 
         setTabs(tabs);