+++ core/source/org/libreoffice/ide/eclipse/core/launch/office/OfficeLaunchDelegate.java
@@ -55,6 +55,7 @@ import org.libreoffice.ide.eclipse.core.model.IUnoidlProject;
 import org.libreoffice.ide.eclipse.core.model.ProjectsManager;
 import org.libreoffice.ide.eclipse.core.model.config.IOOo;
 import org.libreoffice.ide.eclipse.core.model.config.NullExtraOptionsProvider;
+import org.libreoffice.ide.eclipse.core.model.pack.PackagePropertiesModel;
 import org.libreoffice.plugin.core.model.UnoPackage;
 
 /**
@@ -106,7 +107,8 @@ public class OfficeLaunchDelegate extends LaunchConfigurationDelegate {
                     IProject prj = ResourcesPlugin.getWorkspace().getRoot().getProject(prjName);
                     TypesBuilder.build(prj, pMonitor);
 
-                    List<IResource> resources = PackageConfigTab.getResources(pConfiguration);
+                    PackagePropertiesModel pPropertiesModel = new PackagePropertiesModel(prj.getFile("package.properties"));
+                    List<IResource> resources = pPropertiesModel.getContents();
                     File destFile = exportComponent(unoprj, resources);
                     pMonitor.worked(1);