+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Support for LibreOffice 5.0 added. Older releases are no longer supported.
 * Support for Eclipse Mars (4.5) added.
 * Deployment: Install extensions without asking when in debug mode or using a separate user profile
+* Runtime configuration now reads package contents from plugin.properties instead of having an own tab for this
 * Renamed the extension and internal paths (OpenOffice -> LibreOffice)
 * Hosting now by [The Document Foundation](http://www.documentfoundation.org/)
 * Java 7 is the new baseline for the code (instead of Java 5)